1.75.3 (5.6.4f1)
Build-Id: 6b237e17-95fc-47d3-8ba2-010877881f1c
Build-Sha256: baadd4baed86312bd2785a8d15e313df2eb952a965a6f0e48e111fd12207a280

+++ Assembly-CSharp/DesktopCubeModelingToolsController.cs
@@ -58,7 +58,7 @@ public class DesktopCubeModelingToolsController : MonoBehaviour
 
 	private void Start()
 	{
-		ActiveTool = (CubeModelingEvent)(int)cubeModelingStateMachine.curEvent;
+		ActiveTool = (CubeModelingEvent)cubeModelingStateMachine.curEvent;
 		if (defaultTool == null)
 		{
 			SetButtonTransparency(ActiveTool);
